Wire up DartGlobal.* generation again.
We removed this in a previous commit, but I brought
it back here because we want the autogenerated symbolizer
resolver pair.  I moved from WebKitClassIds.* to DartGlobal.*
to match what the template was named, but we can find a better
name for these in the future.

TBR=abarth@chromium.org,terry@google.com
BUG=

Review URL: https://codereview.chromium.org/908663002