Sky should be able to load Dart libraries
This CL teaches Sky how to load Dart modules. It implements a DartLoader and a
mechanism for watching dependency resolution.

We don't block the parser while we're waiting for Dart libraries to load. However,
we'll likely want to do that in the future in order to make sure we register
elements before continuing parsing.

Review URL: https://codereview.chromium.org/910713003